--- src/components/common/createBtn.jsx
@@ -1,11 +1,0 @@
-import { StyledButton } from "./index.style";
-
-const CreateBtn = (onclick) => {
-    return(
-        <StyledButton onclick={onclick}>
-            생성하기
-        </StyledButton>
-    );
-};
-
-export default CreateBtn
