+++ src/components/createBtn/CreateBtn.jsx
@@ -1,0 +1,11 @@
+import { StyledButton } from "./index.style";
+
+const CreateBtn = (onclick) => {
+    return(
+        <StyledButton onclick={onclick}>
+            생성하기
+        </StyledButton>
+    );
+};
+
+export default CreateBtn
